# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ROBOTFORGE API
  description: Robot Data Infrastructure for Physical AI
  version: 0.1.0
  contact:
    name: Mohammed Maheer
    email: maheer@accelerationrobotics.com
servers:
  - url: https://api.robotforge.io/v1
    description: Production
  - url: http://localhost:3000/api
    description: Local development
paths:
...
  /datasets:
    get:
      tags: [Marketplace]
      summary: Search and list datasets
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: task
          in: query
          schema: { type: string }
        - name: embodiment
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Dataset list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dataset'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Marketplace]
      summary: Create dataset listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetCreate'
      responses:
        '201':
          description: Dataset created
  /datasets/{id}:
    get:
      tags: [Marketplace]
      summary: Get dataset detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /datasets/{id}/purchase:
    post:
      tags: [Marketplace]
      summary: Purchase dataset access
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Purchase successful

  # ── Streaming (WebRTC) ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Dataset:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
        task: { type: string }
        embodiments: { type: array, items: { type: string } }
        episodeCount: { type: integer }
        totalDurationHours: { type: number }
        sizeGb: { type: number }
        qualityScore: { type: number }
        format: { type: string }
        pricingTier: { type: string }
        pricePerEpisode: { type: number }
        tags: { type: array, items: { type: string } }
        downloads: { type: integer }
        rating: { type: number }
        createdAt: { type: string, format: date-time }
    DatasetCreate:
      type: object
      required: [name, description, task, embodiments, format]
      properties:
        name: { type: string }
        description: { type: string }
        task: { type: string }
        embodiments: { type: array, items: { type: string } }
        format: { type: string }
        pricingTier: { type: string }
        pricePerEpisode: { type: number }
        tags: { type: array, items: { type: string } }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Fleet Schemas ──────────────────────────────────────
...